+++ .storybook/config.js
@@ -1,7 +1,13 @@
 import { configure } from '@storybook/react'
+import { getStorybook } from '@storybook/react'
+import 'react-chromatic/storybook-addon'
+
+let req = require.context('../src/components/', true, /story\.js$/)
+if (process.env.SKETCH) req = require.context('../src/components/', true, /sketch\.js$/)
 
 function loadStories() {
-  require('../src/stories')
+  req.keys().forEach(req)
 }
 
 configure(loadStories, module)
+export { getStorybook }