+++ src/components/babel.js
@@ -1,3 +1,0 @@
-const config = require('babel-preset-cosmos')
-
-module.exports = config