+++ src/components/molecules/form/field/stories/textarea-field.story.js
@@ -1,0 +1,26 @@
+import React from 'react'
+import { storiesOf } from '@storybook/react'
+import { Example, Stack } from '@auth0/cosmos/_helpers/story-helpers'
+
+import { Form } from '@auth0/cosmos'
+
+storiesOf('Form').add('textarea field', () => (
+  <Example title="textarea field">
+    <Form>
+      <Form.TextArea label="Field label" type="text" placeholder="Enter something" />
+    </Form>
+  </Example>
+))
+
+storiesOf('Form').add('textarea field + error', () => (
+  <Example title="textarea field + error">
+    <Form>
+      <Form.TextArea
+        label="Field label"
+        error="Everything is broken"
+        type="text"
+        placeholder="Enter something"
+      />
+    </Form>
+  </Example>
+))