+++ src/docs/spec/example.js
@@ -6,11 +6,10 @@ import Playground from './playground'
 import Break from './break'
 
 import IconBrowser from './icon-browser'
-import { Code } from '@auth0/cosmos'
+import { Alert, Code } from '@auth0/cosmos'
 import { Text, ListItem, List, Link } from '../docs-components/typography'
 import PageHeader from './page-header'
-import SectionHeader from './section-header'
-import ExampleHeader from './example-header'
+import { SectionHeader, ExampleHeader } from './header'
 
 const Example = props => {
   const options = {
@@ -44,6 +43,7 @@ const Example = props => {
           return null
         }
       },
+      Alert,
       IconBrowser
     }
   }