+++ src/overview/index.js
@@ -16,12 +16,14 @@ import Tooltip from './examples/tooltip'
 import Inputs from './examples/inputs'
 import Typography from './examples/typography'
 import TabsExample from './examples/tabs'
-import Lists from './examples/lists'
 import Buttons from './examples/buttons'
 import ButtonsInputs from './examples/buttons-inputs'
 import Logos from './examples/logos'
 import Avatars from './examples/avatars'
 import Thumbnails from './examples/thumbnails'
+import Lists from './examples/lists'
+import ResourceLists from './examples/resource-lists'
+import Tables from './examples/tables'
 
 export default () => (
   <Overview>
@@ -50,6 +52,14 @@ export default () => (
       </Row>
       <Row>
         <Col>
+          <Avatars />
+        </Col>
+        <Col>
+          <Thumbnails />
+        </Col>
+      </Row>
+      <Row>
+        <Col>
           <Tooltip />
         </Col>
         <Col>
@@ -60,14 +70,8 @@ export default () => (
       <TabsExample />
       <EmptyStates />
       <Lists />
-      <Row>
-        <Col>
-          <Avatars />
-        </Col>
-        <Col>
-          <Thumbnails />
-        </Col>
-      </Row>
+      <ResourceLists />
+      <Tables />
     </Container>
   </Overview>
 )